requirements.txt
boltons==24.1.0
hyperlink==21.0.0
pyyaml==6.0.2
schema==0.7.7
